Nothing implicit: the code says what it means, and the analysis reads every file of it
Eight checks join the gate, none of them about a fault and all of them about saying what the code means, which is this project's rule for its own writing. readability-implicit-bool-conversion has a pointer compared with nullptr and a number with nought rather than standing in for a truth of its own. misc-const-correctness has a local that is never written say const, so that the names a piece of code can still change are the ones it does not mark: two hundred and sixty lines say it now, fifty-eight of them declarations of four names at once, which the check reports and cannot fix since the const has to go in front of all four. cppcoreguidelines-init-variables has a local hold a value from the line it is declared on, which four places did not: the kind of a single-instrument file and the least it can hold, the step of the level monitor, and the fine tuning of the second voice in each direction. Its own fix would put nought there, which says the value is nought when it is not, so each of them instead works the value out where the name appears -- and is const as well. readability-named-parameter has every parameter say what it is in the declaration as well as the definition, with the names of the ones a function does not use in comments, since such a name is a warning of its own. The other four -- cppcoreguidelines-narrowing-conversions, google-explicit-constructor, readability-qualified-auto and readability-static-accessed-through-instance -- say nothing about this project today, and are here so that the next line which does otherwise is answered as it is written.

Where a check is wrong it is answered where it is wrong, with the reason. misc-const-correctness offers a pointee that is const in eight places; in three of them the pointee is written through a copy of the pointer, which it does not see: the samples of a block are written by the processor the buffer is handed to, and the operator editors are set through the pointers of an array. It offers one more for a loop that deletes the dialogs it walks, which the language allows through a pointer to const and which is the largest change there is, and one for the function that clears the marks of blank instruments, whose whole purpose is to change the file it is given. The parameters of the handler Xlib is given are Xlib's, and a pointee that is const would be a different type and no longer that function.

Two places had a const the check could see on one chip and not the other, since what writes the variable sits under an #if. Both are written so that the const holds on either: the format a single instrument file is loaded in is settled per chip where it is declared, and the fuzz target that compares an instrument with itself saved and loaded again clears the mark of a blank instrument where it makes each side rather than afterwards.

The analysis now reads every file of ours, which it did not. No file of this project is a C++ module and none imports one, and CMake was scanning every C++23 source for them and naming the mapping file it writes in the compile command: a reader of the code could not parse a single one of those files until a build had generated it, which is what the job of the last commit failed on. The scanning is off. The job no longer configures the fuzz targets either, which it did so that their sources were read: a build with the tests carries the program that replays a saved input through each target, and that compiles every one of those sources, while the coverage those targets ask of the libraries is what kept the tool that writes the pack of banks from linking -- a build that instruments the libraries cannot link a program that brings no sanitizer runtime. With the tools in the build for the first time, tools/render had six things to answer: std::endl, where the flush is wanted and is now written out; three loops whose induction variable was a double, now counted by an integer step whose time is worked out from the count, so that how many steps there are is not a matter of what a sum of doubles comes to; and two appends that built a string to move it.

-Wzero-as-null-pointer-constant is asked of the C as well as the C++ now. The C of this project is C23 and has nullptr, so nought standing in for it is as much a thing this project does not write there; clang has nothing to say about it in C today, and clang-tidy's modernize-use-nullptr is what catches it. A check that repeats a warning is kept for the same reason: it costs nothing, and it holds where the warning might not reach.

+++ sources/opn2/adl/instrument.cc
@@ -86,7 +86,7 @@ void Instrument::describe(std::FILE *out) const noexcept
                  " - AM sensitivity %d FM sensitivity %d\n"
                  " - Velocity offset %d\n"
                  " - Percussion note %d\n",
-                 blank(),
+                 static_cast<int>(blank()),
                  feedback(), algorithm(), note_offset,
                  ams(), fms(),
                  midi_velocity_offset, percussion_key_number);
@@ -103,8 +103,8 @@ void Instrument::describe_operator(unsigned op, std::FILE *out, const char *inde
                  "%s - SSG-EG Enable %d Wave %d\n",
                  indent, op,
                  indent, attack(op), decay2(op), decay1(op), sustain(op), release(op),
-                 indent, am(op), level(op), ratescale(op), detune(op), fmul(op),
-                 indent, ssgenable(op), ssgwave(op));
+                 indent, static_cast<int>(am(op)), level(op), ratescale(op), detune(op), fmul(op),
+                 indent, static_cast<int>(ssgenable(op)), ssgwave(op));
 }
 
 // Compares field by field. The structs have padding, so memcmp could tell
